Fix Part of #10700: Refactor Topic Object (#16805)
* TopicObjectFactory is migrated

* fixing linter check 2
