+++ core/templates/pages/skill-editor-page/editor-tab/skill-editor-main-tab.component.ts
@@ -22,7 +22,7 @@ import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { SavePendingChangesModalComponent } from 'components/save-pending-changes/save-pending-changes-modal.component';
 import { UndoRedoService } from 'domain/editor/undo_redo/undo-redo.service';
 import { Skill } from 'domain/skill/SkillObjectFactory';
-import { Topic } from 'domain/topic/TopicObjectFactory';
+import { Topic } from 'domain/topic/topic-object.model';
 import { PageTitleService } from 'services/page-title.service';
 import { FocusManagerService } from 'services/stateful/focus-manager.service';
 import { SkillEditorRoutingService } from '../services/skill-editor-routing.service';
