+++ core/templates/pages/topic-editor-page/editor-tab/topic-editor-tab.directive.spec.ts
@@ -27,7 +27,7 @@ import { StoryReference } from 'domain/topic/story-reference-object.model';
 // TODO(#7222): Remove the following block of unnnecessary imports once
 // the code corresponding to the spec is upgraded to Angular 8.
 import { importAllAngularServices } from 'tests/unit-test-utils.ajs';
-import { Topic } from 'domain/topic/TopicObjectFactory';
+import { Topic } from 'domain/topic/topic-object.model';
 // ^^^ This block is to be removed.
 
 describe('Topic editor tab directive', function() {
