+++ core/templates/pages/topic-editor-page/services/subtopic-validation.service.spec.ts
@@ -19,7 +19,7 @@
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { TestBed, waitForAsync } from '@angular/core/testing';
 import { Subtopic } from 'domain/topic/subtopic.model';
-import { Topic, TopicObjectFactory } from 'domain/topic/TopicObjectFactory';
+import { Topic } from 'domain/topic/topic-object.model';
 import { SubtopicValidationService } from './subtopic-validation.service';
 import { TopicEditorStateService } from './topic-editor-state.service';
 
@@ -34,7 +34,6 @@ describe('Subtopic validation service', () => {
       ],
       providers: [
         TopicEditorStateService,
-        TopicObjectFactory
       ]
     }).compileComponents();
   }));
