+++ core/templates/pages/topic-editor-page/services/topic-editor-state.service.spec.ts
@@ -25,7 +25,7 @@ import { EditableTopicBackendApiService, FetchTopicResponse, UpdateTopicResponse
 import { SubtopicPage, SubtopicPageBackendDict } from 'domain/topic/subtopic-page.model';
 import { TopicRightsBackendApiService } from 'domain/topic/topic-rights-backend-api.service';
 import { TopicRights, TopicRightsBackendDict } from 'domain/topic/topic-rights.model';
-import { TopicBackendDict, TopicObjectFactory } from 'domain/topic/TopicObjectFactory';
+import { TopicBackendDict, Topic } from 'domain/topic/topic-object.model';
 import { AlertsService } from 'services/alerts.service';
 import { TopicEditorStateService } from './topic-editor-state.service';
 
@@ -34,7 +34,6 @@ describe('Topic editor state service', () => {
   let mockEditableTopicBackendApiService: MockEditableTopicBackendApiService;
   let alertsService: AlertsService;
   let undoRedoService: UndoRedoService;
-  let topicObjectFactory: TopicObjectFactory;
   let editableStoryBackendApiService: EditableStoryBackendApiService;
 
   let skillCreationIsAllowed: boolean = true;
@@ -162,7 +161,6 @@ describe('Topic editor state service', () => {
           provide: EditableTopicBackendApiService,
           useClass: MockEditableTopicBackendApiService
         },
-        TopicObjectFactory,
         {
           provide: TopicRightsBackendApiService,
           useClass: MockTopicRightsBackendApiService
@@ -182,7 +180,6 @@ describe('Topic editor state service', () => {
       TestBed.inject(EditableStoryBackendApiService);
     alertsService = (TestBed.inject(AlertsService) as unknown) as
       jasmine.SpyObj<AlertsService>;
-    topicObjectFactory = TestBed.inject(TopicObjectFactory);
 
     topicDict = {
       id: 'topic_id',
@@ -332,14 +329,14 @@ describe('Topic editor state service', () => {
   }));
 
   it('should set topic', () => {
-    let topic = topicObjectFactory.create(topicDict, {});
+    let topic = Topic.create(topicDict, {});
     topicEditorStateService.setTopic(topic);
     expect(topicEditorStateService.getTopic()).toEqual(topic);
   });
 
   it('should delete subtopic page when user deletes subtopic', fakeAsync(() => {
     topicDict.id = 'topic_id1234';
-    let topic = topicObjectFactory.create(topicDict, {});
+    let topic = Topic.create(topicDict, {});
     topicEditorStateService.setTopic(topic);
     subtopicPage.id = 'topic_id1234-0';
     topicEditorStateService.setSubtopicPage(
@@ -377,7 +374,7 @@ describe('Topic editor state service', () => {
     spyOn(editableStoryBackendApiService, 'deleteStoryAsync').and
       .callThrough();
     var successCallback = jasmine.createSpy('successCallback');
-    let topic = topicObjectFactory.create(topicDict, {});
+    let topic = Topic.create(topicDict, {});
     topicEditorStateService.setTopic(topic);
 
     topicEditorStateService.saveTopic('Commit Message', successCallback);
@@ -410,7 +407,7 @@ describe('Topic editor state service', () => {
         .returnValue(Promise.reject());
       spyOn(alertsService, 'addWarning');
       var successCallback = jasmine.createSpy('successCallback');
-      let topic = topicObjectFactory.create(topicDict, {});
+      let topic = Topic.create(topicDict, {});
       topicEditorStateService.setTopic(topic);
 
       topicEditorStateService.saveTopic('Commit Message', successCallback);
@@ -434,7 +431,7 @@ describe('Topic editor state service', () => {
     spyOn(undoRedoService, 'hasChanges').and.returnValue(false);
     spyOn(mockEditableTopicBackendApiService, 'updateTopicAsync');
 
-    let topic = topicObjectFactory.create(topicDict, {});
+    let topic = Topic.create(topicDict, {});
     topicEditorStateService.setTopic(topic);
 
     topicEditorStateService.saveTopic('Commit Message', () => {});
@@ -459,7 +456,7 @@ describe('Topic editor state service', () => {
   }));
 
   it('should set topic rights when called', () => {
-    let topic = topicObjectFactory.create(topicDict, {});
+    let topic = Topic.create(topicDict, {});
     topicEditorStateService.setTopic(topic);
     expect(topicEditorStateService.getTopicRights()).toEqual(
       TopicRights.createFromBackendDict({
